chore(configuration): vanilla extract config
