+++ app/components/button/index.tsx
@@ -1,6 +1,5 @@
 import * as styles from "./styles.css";
 import React from "react";
-import { useThemeProvider } from "~/context"; // Note that `.ts` is omitted here
 
 interface ButtonProps {
   children: React.ReactNode;
