+++ app/configs/theme/contract.css.ts
@@ -1,0 +1,4 @@
+import { createThemeContract } from "@vanilla-extract/css";
+import { THEME_CONTRACT } from "~/configs/theme/default-theme";
+
+export const themeVars = createThemeContract(THEME_CONTRACT);
