--- app/configs/theme/contract.ts
@@ -1,4 +1,0 @@
-import { createThemeContract } from "@vanilla-extract/css";
-import { DEFAULT_THEMES } from "~/configs/theme/default-theme";
-
-export const themeVars = createThemeContract(DEFAULT_THEMES);
