--- app/configs/theme/index.ts
@@ -1,2 +1,0 @@
-export * from "./theme.css";
-export * from "./contract";
