+++ app/configs/theme/theme.css.ts
@@ -1,12 +1,5 @@
 import { createTheme } from "@vanilla-extract/css";
-import {
-  CARNATION,
-  CORAL,
-  LAVENDER,
-  SHOCKING_PINK,
-  SUNSHADE,
-} from "~/configs/theme/colors";
-import type { ThemeObject } from "~/configs/theme/type";
 import { DEFAULT_THEMES } from "~/configs/theme/default-theme";
+import { themeVars } from "~/configs/theme/contract.css";
 
-export const theme = createTheme<ThemeObject>(DEFAULT_THEMES);
+export const theme = createTheme(themeVars, DEFAULT_THEMES);
