+++ app/context/theme.context.ts
@@ -1,14 +1,9 @@
 import React from "react";
-import type { ThemeObject } from "~/configs/theme/type";
-import type { CSSVarFunction, MapLeafNodes } from "@vanilla-extract/private";
-import { theme } from "~/configs/theme";
+import { theme } from "~/configs/theme/theme.css";
 
 interface ThemeContext {
   currentTheme: string;
-  themes: Record<
-    string,
-    [className: string, vars: MapLeafNodes<ThemeObject, CSSVarFunction>]
-  >;
+  themes: Record<string, string>;
   onThemeChange?: (theme: string) => void;
 }
