+++ package.json
@@ -14,15 +14,15 @@
     "@remix-run/node": "^2.2.0",
     "@remix-run/react": "^2.2.0",
     "@remix-run/serve": "^2.2.0",
-    "@vanilla-extract/css": "^1.14.0",
-    "@vanilla-extract/css-utils": "^0.1.3",
-    "@vanilla-extract/dynamic": "^2.1.0",
-    "@vanilla-extract/private": "^1.0.3",
     "isbot": "^3.6.8",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
+    "@vanilla-extract/css": "^1.14.0",
+    "@vanilla-extract/css-utils": "^0.1.3",
+    "@vanilla-extract/dynamic": "^2.1.0",
+    "@vanilla-extract/private": "^1.0.3",
     "@remix-run/dev": "^2.2.0",
     "@remix-run/eslint-config": "^2.2.0",
     "@types/react": "^18.2.20",
